Simulated EU-SILC data - aggregated population data

The data set includes synthetic EU-SILC data and is taken from the package emdi. Originally, the data builds on eusilcP from package simFrame and is reduced to 15 variables including district identifiers as well as aggregated household level covariates. Therefore, except for the variables ratio_n and district, the variables are mean values per district.

Usage

eusilcA_popAgg

Format

A data frame with 94 observations and 15 variables:

eqsize

numeric; the equivalized household size according to the modified OECD scale.

cash

numeric; employee cash or near cash income (net).

self_empl

numeric; cash benefits or losses from self-employment (net).

unempl_ben

numeric; unemployment benefits (net).

age_ben

numeric; old-age benefits (net).

surv_ben

numeric; survivor's benefits (net).

sick_ben

numeric; sickness benefits (net).

dis_ben

numeric; disability benefits (net).

rent

numeric; income from rental of a property or land (net).

fam_allow

numeric; family/children related allowances (net).

house_allow

numeric; housing allowances (net).

cap_inv

numeric; interest, dividends, profit from capital investments in unincorporated business (net).

tax_adj

numeric; repayments/receipts for tax adjustment (net).

ratio_n

numeric; ratios of the population size per area and the total population size.

district

factor; Austrian districts (94 levels).