Presents point, MSE and CV estimates

Function summarize_indicators reports point and mean squared error (MSE) estimates as well as calculated coefficients of variation (CV) from a fitted SAEforest object.

Usage

summarize_indicators(object, indicator = "all", MSE = FALSE, CV = FALSE)

Arguments

object

Object for which point and/or MSE estimates and/or calculated CV's are requested. The object must be of class SAEforest.

indicator

Optional character vector specifying indicators to be mapped: (i) all calculated indicators ("all"); (ii) each default indicators name: "Mean", "Quant10", "Quant25", "Median", "Quant75", "Quant90", "Gini", "Hcr", "Pgap", "Qsr" or the function name/s of "custom_indicator/s"; (iii) a vector of names of indicators. If the object is estimated by SAEforest_model indicator arguments are ignored and only the "Mean" is returned.

MSE

Logical. If TRUE, MSE estimates for selected indicators per domain are added to the data frame of point estimates. Defaults to FALSE.

CV

Logical. If TRUE, coefficients of variation for selected indicators per domain are added to the data frame of point estimates. Defaults to FALSE.

Value

The return of summarize_indicators is an object of class summarize_indicators.SAEforest

including domain-specific point and/or MSE estimates and/or calculated CV's from a SAEforest object The returned object contains the data.frame ind and a character including the names of requested indicator(s).

Details

Objects of class summarize_indicators.SAEforest have methods for following generic functions: head and tail (for default documentation, see head), as.matrix (for default documentation, see matrix), as.data.frame (for default documentation, see as.data.frame), subset (for default documentation, see subset).

See also

Examples

# \donttest{
# Loading data
data("eusilcA_pop")
data("eusilcA_smp")

income <- eusilcA_smp$eqIncome
X_covar <- eusilcA_smp[, -c(1, 16, 17, 18)]

# Calculating point + MSE estimates and passing arguments to the forest.
# Additionally, two additional indicators and functions as threshold are added.
# Note that B and num.trees are low to speed up estimation time and must be changed for
# practical applications.

model1 <- SAEforest_model(Y = income, X = X_covar, dName = "district",
                          smp_data = eusilcA_smp, pop_data = eusilcA_pop,
                          meanOnly = FALSE, MSE = "nonparametric", B = 5, mtry = 5,
                          num.trees = 50, smearing = FALSE)
#> Error SD Bootstrap started:
#> MSE Bootstrap with 5 rounds started:

# Extract indicator and show generics:
Gini1 <- summarize_indicators(model1, MSE = TRUE, CV = TRUE, indicator = "Gini")

head(Gini1)
#>              district      Gini     Gini_MSE    Gini_CV
#> 1           Amstetten 0.2506610 0.0004543057 0.08503297
#> 2               Baden 0.1815458 0.0001324130 0.06338390
#> 3             Bludenz 0.2841038 0.0007885591 0.09884166
#> 4      Braunau am Inn 0.2697336 0.0002901931 0.06315512
#> 5             Bregenz 0.1617372 0.0001258838 0.06937054
#> 6 Bruck an der Leitha 0.1778989 0.0002382211 0.08675945
tail(Gini1)
#>                   district      Gini     Gini_MSE    Gini_CV
#> 89                    Wien 0.2413731 0.0001131947 0.04407824
#> 90  Wiener Neustadt (Land) 0.2006023 0.0002253390 0.07483112
#> 91 Wiener Neustadt (Stadt) 0.2100767 0.0014837030 0.18335616
#> 92               Wolfsberg 0.2311071 0.0009660102 0.13448613
#> 93             Zell am See 0.3003195 0.0004827088 0.07315753
#> 94                  Zwettl 0.2677981 0.0022288514 0.17629216
as.data.frame(Gini1)
#>                         district      Gini     Gini_MSE    Gini_CV
#> 1                      Amstetten 0.2506610 4.543057e-04 0.08503297
#> 2                          Baden 0.1815458 1.324130e-04 0.06338390
#> 3                        Bludenz 0.2841038 7.885591e-04 0.09884166
#> 4                 Braunau am Inn 0.2697336 2.901931e-04 0.06315512
#> 5                        Bregenz 0.1617372 1.258838e-04 0.06937054
#> 6            Bruck an der Leitha 0.1778989 2.382211e-04 0.08675945
#> 7             Bruck-Mürzzuschlag 0.1682187 1.652236e-04 0.07641205
#> 8               Deutschlandsberg 0.1913224 2.293643e-04 0.07915843
#> 9                       Dornbirn 0.2082983 6.929718e-04 0.12637817
#> 10                      Eferding 0.2329806 6.966463e-04 0.11328867
#> 11            Eisenstadt (Stadt) 0.1875840 8.119365e-04 0.15190262
#> 12           Eisenstadt-Umgebung 0.2004414 5.140637e-04 0.11311524
#> 13                     Feldkirch 0.2262377 5.047038e-04 0.09930092
#> 14                   Feldkirchen 0.2537076 1.579048e-03 0.15662611
#> 15                     Freistadt 0.2099376 8.011244e-04 0.13482167
#> 16                       Gmunden 0.2035652 2.554946e-04 0.07852126
#> 17                         Gmünd 0.2557238 2.125490e-03 0.18028451
#> 18                  Graz (Stadt) 0.2155064 1.494586e-04 0.05672838
#> 19                 Graz-Umgebung 0.1829827 1.316544e-03 0.19829311
#> 20                  Grieskirchen 0.2090666 1.008586e-03 0.15190489
#> 21                   Gänserndorf 0.1994797 1.921915e-04 0.06949737
#> 22                       Güssing 0.2301604 7.912222e-04 0.12221336
#> 23                       Hallein 0.2254833 6.489844e-04 0.11298030
#> 24          Hartberg-Fürstenfeld 0.2631672 7.679258e-04 0.10529988
#> 25                      Hermagor 0.2485312 1.650781e-03 0.16347970
#> 26                    Hollabrunn 0.2197322 4.093010e-04 0.09207203
#> 27                          Horn 0.2289698 3.579521e-04 0.08262932
#> 28                          Imst 0.2449579 7.311943e-04 0.11038880
#> 29              Innsbruck (Land) 0.2019746 4.139933e-04 0.10073955
#> 30             Innsbruck (Stadt) 0.2069875 3.551326e-04 0.09104397
#> 31                   Jennersdorf 0.2459114 5.877856e-04 0.09858954
#> 32        Kirchdorf an der Krems 0.2302639 8.306758e-04 0.12516700
#> 33                     Kitzbühel 0.2523222 1.222655e-03 0.13857873
#> 34             Klagenfurt (Land) 0.1815504 5.944269e-04 0.13429261
#> 35            Klagenfurt (Stadt) 0.1627702 4.077377e-04 0.12405537
#> 36                    Korneuburg 0.1608567 9.234021e-05 0.05973879
#> 37                  Krems (Land) 0.2358282 5.561211e-04 0.09999745
#> 38    Krems an der Donau (Stadt) 0.2181051 7.553568e-04 0.12601150
#> 39                      Kufstein 0.1833995 1.678502e-04 0.07064198
#> 40                       Landeck 0.2602467 1.148258e-03 0.13020708
#> 41                      Leibnitz 0.2350626 3.335657e-04 0.07769751
#> 42                        Leoben 0.1860993 1.033475e-04 0.05462673
#> 43                         Lienz 0.2627103 6.914913e-04 0.10009588
#> 44                        Liezen 0.2587971 9.768290e-04 0.12076743
#> 45                    Lilienfeld 0.2417432 6.898991e-04 0.10865219
#> 46                  Linz (Stadt) 0.1847478 1.815338e-04 0.07292888
#> 47                     Linz-Land 0.1684694 3.443728e-05 0.03483321
#> 48                   Mattersburg 0.2073792 7.126628e-04 0.12872916
#> 49                          Melk 0.2656890 1.255416e-03 0.13335838
#> 50                    Mistelbach 0.2002418 2.197294e-04 0.07402685
#> 51                         Murau 0.2544567 2.923920e-03 0.21250489
#> 52                        Murtal 0.2232718 3.604960e-04 0.08503863
#> 53                       Mödling 0.1454743 7.551740e-04 0.18890225
#> 54                   Neunkirchen 0.2177180 2.832909e-04 0.07730757
#> 55               Neusiedl am See 0.2075293 5.246292e-04 0.11036894
#> 56                Oberpullendorf 0.2225282 5.380646e-04 0.10423948
#> 57                      Oberwart 0.2521503 4.386798e-04 0.08306428
#> 58                          Perg 0.2072744 2.015107e-04 0.06848623
#> 59                        Reutte 0.2070370 3.879181e-04 0.09513098
#> 60              Ried im Innkreis 0.2438417 2.854332e-04 0.06928579
#> 61                      Rohrbach 0.2279989 9.168689e-04 0.13280699
#> 62                  Rust (Stadt) 0.2508950 4.058043e-03 0.25390217
#> 63              Salzburg (Stadt) 0.1919434 1.643715e-04 0.06679440
#> 64             Salzburg-Umgebung 0.1639323 1.139990e-04 0.06513076
#> 65        Sankt Johann im Pongau 0.2402848 4.451068e-04 0.08780229
#> 66           Sankt Pölten (Land) 0.2332269 1.503084e-04 0.05256697
#> 67          Sankt Pölten (Stadt) 0.2199186 3.331637e-04 0.08299787
#> 68        Sankt Veit an der Glan 0.2438201 8.221905e-04 0.11760254
#> 69                      Scheibbs 0.2442704 5.390493e-03 0.30056843
#> 70                        Schwaz 0.2317185 4.676871e-04 0.09332906
#> 71                     Schärding 0.2907355 5.480095e-04 0.08051855
#> 72           Spittal an der Drau 0.2521698 6.586615e-04 0.10177429
#> 73                 Steyr (Stadt) 0.2102186 1.969636e-03 0.21111638
#> 74                    Steyr-Land 0.1731385 6.047411e-04 0.14203360
#> 75              Südoststeiermark 0.2439255 9.648467e-04 0.12734208
#> 76                       Tamsweg 0.2094321 1.302228e-04 0.05448792
#> 77                         Tulln 0.1996404 6.394416e-05 0.04005457
#> 78               Urfahr-Umgebung 0.1420431 7.835259e-04 0.19706365
#> 79               Villach (Stadt) 0.1877881 6.452013e-05 0.04277396
#> 80                  Villach Land 0.2042127 3.694955e-04 0.09412866
#> 81                     Voitsberg 0.2111620 1.451617e-03 0.18043062
#> 82                   Vöcklabruck 0.2094107 5.244822e-04 0.10936199
#> 83                   Völkermarkt 0.2369843 7.105056e-04 0.11247713
#> 84        Waidhofen an der Thaya 0.2672654 3.274056e-03 0.21409191
#> 85 Waidhofen an der Ybbs (Stadt) 0.2213196 3.763664e-03 0.27719509
#> 86                          Weiz 0.2436176 2.927377e-04 0.07023130
#> 87                  Wels (Stadt) 0.2369898 1.084322e-03 0.13894711
#> 88                     Wels-Land 0.1967274 7.877112e-04 0.14266538
#> 89                          Wien 0.2413731 1.131947e-04 0.04407824
#> 90        Wiener Neustadt (Land) 0.2006023 2.253390e-04 0.07483112
#> 91       Wiener Neustadt (Stadt) 0.2100767 1.483703e-03 0.18335616
#> 92                     Wolfsberg 0.2311071 9.660102e-04 0.13448613
#> 93                   Zell am See 0.3003195 4.827088e-04 0.07315753
#> 94                        Zwettl 0.2677981 2.228851e-03 0.17629216
as.matrix(Gini1)
#>            Gini     Gini_MSE    Gini_CV
#>  [1,] 0.2506610 4.543057e-04 0.08503297
#>  [2,] 0.1815458 1.324130e-04 0.06338390
#>  [3,] 0.2841038 7.885591e-04 0.09884166
#>  [4,] 0.2697336 2.901931e-04 0.06315512
#>  [5,] 0.1617372 1.258838e-04 0.06937054
#>  [6,] 0.1778989 2.382211e-04 0.08675945
#>  [7,] 0.1682187 1.652236e-04 0.07641205
#>  [8,] 0.1913224 2.293643e-04 0.07915843
#>  [9,] 0.2082983 6.929718e-04 0.12637817
#> [10,] 0.2329806 6.966463e-04 0.11328867
#> [11,] 0.1875840 8.119365e-04 0.15190262
#> [12,] 0.2004414 5.140637e-04 0.11311524
#> [13,] 0.2262377 5.047038e-04 0.09930092
#> [14,] 0.2537076 1.579048e-03 0.15662611
#> [15,] 0.2099376 8.011244e-04 0.13482167
#> [16,] 0.2035652 2.554946e-04 0.07852126
#> [17,] 0.2557238 2.125490e-03 0.18028451
#> [18,] 0.2155064 1.494586e-04 0.05672838
#> [19,] 0.1829827 1.316544e-03 0.19829311
#> [20,] 0.2090666 1.008586e-03 0.15190489
#> [21,] 0.1994797 1.921915e-04 0.06949737
#> [22,] 0.2301604 7.912222e-04 0.12221336
#> [23,] 0.2254833 6.489844e-04 0.11298030
#> [24,] 0.2631672 7.679258e-04 0.10529988
#> [25,] 0.2485312 1.650781e-03 0.16347970
#> [26,] 0.2197322 4.093010e-04 0.09207203
#> [27,] 0.2289698 3.579521e-04 0.08262932
#> [28,] 0.2449579 7.311943e-04 0.11038880
#> [29,] 0.2019746 4.139933e-04 0.10073955
#> [30,] 0.2069875 3.551326e-04 0.09104397
#> [31,] 0.2459114 5.877856e-04 0.09858954
#> [32,] 0.2302639 8.306758e-04 0.12516700
#> [33,] 0.2523222 1.222655e-03 0.13857873
#> [34,] 0.1815504 5.944269e-04 0.13429261
#> [35,] 0.1627702 4.077377e-04 0.12405537
#> [36,] 0.1608567 9.234021e-05 0.05973879
#> [37,] 0.2358282 5.561211e-04 0.09999745
#> [38,] 0.2181051 7.553568e-04 0.12601150
#> [39,] 0.1833995 1.678502e-04 0.07064198
#> [40,] 0.2602467 1.148258e-03 0.13020708
#> [41,] 0.2350626 3.335657e-04 0.07769751
#> [42,] 0.1860993 1.033475e-04 0.05462673
#> [43,] 0.2627103 6.914913e-04 0.10009588
#> [44,] 0.2587971 9.768290e-04 0.12076743
#> [45,] 0.2417432 6.898991e-04 0.10865219
#> [46,] 0.1847478 1.815338e-04 0.07292888
#> [47,] 0.1684694 3.443728e-05 0.03483321
#> [48,] 0.2073792 7.126628e-04 0.12872916
#> [49,] 0.2656890 1.255416e-03 0.13335838
#> [50,] 0.2002418 2.197294e-04 0.07402685
#> [51,] 0.2544567 2.923920e-03 0.21250489
#> [52,] 0.2232718 3.604960e-04 0.08503863
#> [53,] 0.1454743 7.551740e-04 0.18890225
#> [54,] 0.2177180 2.832909e-04 0.07730757
#> [55,] 0.2075293 5.246292e-04 0.11036894
#> [56,] 0.2225282 5.380646e-04 0.10423948
#> [57,] 0.2521503 4.386798e-04 0.08306428
#> [58,] 0.2072744 2.015107e-04 0.06848623
#> [59,] 0.2070370 3.879181e-04 0.09513098
#> [60,] 0.2438417 2.854332e-04 0.06928579
#> [61,] 0.2279989 9.168689e-04 0.13280699
#> [62,] 0.2508950 4.058043e-03 0.25390217
#> [63,] 0.1919434 1.643715e-04 0.06679440
#> [64,] 0.1639323 1.139990e-04 0.06513076
#> [65,] 0.2402848 4.451068e-04 0.08780229
#> [66,] 0.2332269 1.503084e-04 0.05256697
#> [67,] 0.2199186 3.331637e-04 0.08299787
#> [68,] 0.2438201 8.221905e-04 0.11760254
#> [69,] 0.2442704 5.390493e-03 0.30056843
#> [70,] 0.2317185 4.676871e-04 0.09332906
#> [71,] 0.2907355 5.480095e-04 0.08051855
#> [72,] 0.2521698 6.586615e-04 0.10177429
#> [73,] 0.2102186 1.969636e-03 0.21111638
#> [74,] 0.1731385 6.047411e-04 0.14203360
#> [75,] 0.2439255 9.648467e-04 0.12734208
#> [76,] 0.2094321 1.302228e-04 0.05448792
#> [77,] 0.1996404 6.394416e-05 0.04005457
#> [78,] 0.1420431 7.835259e-04 0.19706365
#> [79,] 0.1877881 6.452013e-05 0.04277396
#> [80,] 0.2042127 3.694955e-04 0.09412866
#> [81,] 0.2111620 1.451617e-03 0.18043062
#> [82,] 0.2094107 5.244822e-04 0.10936199
#> [83,] 0.2369843 7.105056e-04 0.11247713
#> [84,] 0.2672654 3.274056e-03 0.21409191
#> [85,] 0.2213196 3.763664e-03 0.27719509
#> [86,] 0.2436176 2.927377e-04 0.07023130
#> [87,] 0.2369898 1.084322e-03 0.13894711
#> [88,] 0.1967274 7.877112e-04 0.14266538
#> [89,] 0.2413731 1.131947e-04 0.04407824
#> [90,] 0.2006023 2.253390e-04 0.07483112
#> [91,] 0.2100767 1.483703e-03 0.18335616
#> [92,] 0.2311071 9.660102e-04 0.13448613
#> [93,] 0.3003195 4.827088e-04 0.07315753
#> [94,] 0.2677981 2.228851e-03 0.17629216
subset(Gini1, district == "Wien")
#>    district      Gini     Gini_MSE    Gini_CV
#> 89     Wien 0.2413731 0.0001131947 0.04407824
# }