Main function for the estimation of domain-level (nonlinear) indicators with MERFs

This function enables the use of Mixed Effects Random Forests (MERFs) for applications of Small Area Estimation (SAE). Unit-level survey data on a target and auxiliary covariates is required to produce reliable estimates of various disaggregated economic and inequality indicators. Option meanOnly saves computational time for users that are only interested in the estimation of domain-specific means using unit-level and aggregated auxiliary data. Predefined indicators include the mean, median, quantiles (10%, 25%, 75% and 90%), the head count ratio, the poverty gap, the Gini-coefficient and the quintile share ratio. The MERF algorithm is an algorithmic procedure reminiscent of an EM-algorithm (see Details). Overall, the function serves as a coherent framework for the estimation of point estimates and if requested uncertainty estimates for the indicators. Methodological details are found in Krennmair & Schmid (2022) and Krennmair et al. (2022b). The following examples showcase further potential applications.

Usage

SAEforest_model(
  Y,
  X,
  dName,
  smp_data,
  pop_data,
  MSE = "none",
  meanOnly = TRUE,
  aggData = FALSE,
  smearing = TRUE,
  popnsize = NULL,
  importance = "impurity",
  OOsample_obs = 25,
  ADDsamp_obs = 0,
  w_min = 3,
  B = 100,
  B_adj = 100,
  B_MC = 100,
  threshold = NULL,
  custom_indicator = NULL,
  initialRandomEffects = 0,
  ErrorTolerance = 1e-04,
  MaxIterations = 25,
  na.rm = TRUE,
  ...
)

Arguments

Y

Continuous input value of target variable.

X

Matrix or data.frame of predictive covariates.

dName

Character specifying the name of the domain identifier, for which random intercepts are modeled.

smp_data

data.frame of survey sample data including the specified elements of Y and X.

pop_data

data.frame of unit-level population covariate data X. Please note that the column names of predictive covariates must match column names of smp_data. This holds especially for the name of the domain identifier.

MSE

Character input specifying the type of uncertainty estimates. Available options are: (i) "none" if only point estimates are requested, (ii) "nonparametric" following the MSE bootstrap procedure proposed by Krennmair & Schmid (2022) or by Krennmair et al. (2022a) if aggData = TRUE. (iii) "wild" only for nonlinear indicators proposed by Krennmair et al. (2022b). Defaults to "none".

meanOnly

Logical. Calculating domain-level means only. Defaults to TRUE.

aggData

Logical input indicating whether aggregated covariate information or unit-level covariate information is used for domain-level means. Defaults to FALSE, assuming unit-level covariate data.

smearing

Logical input indicating whether a smearing based approach or a Monte Carlo (MC) version for point estimates should be obtained to estimate (nonlinear) indicators. MC should be used if computational constraints prohibit a smearing approach. For theoretical details see Krennmair et al (2022b). Defaults to TRUE.

popnsize

data.frame, comprising information of population size of domains. Only needed if aggData = TRUE and a MSE is requested. Please note that the name of the domain identifier must match the column name of smp_data.

importance

Variable importance mode processed by the random forest from ranger. Must be 'none', 'impurity', 'impurity_corrected' or 'permutation'. A concept of variable importance is needed for the production of generic plots plot. For the estimation of domain-level means under aggregated covariate data, variable importance is needed to rank information in the process of finding suitable calibration weights (Krennmair et al., 2022b). For further information regarding measures of importance see ranger.

OOsample_obs

Number of out-of-sample observations taken from the closest area for potentially unsampled areas. Only needed if aggData = TRUE with defaults to 25.

ADDsamp_obs

Number of out-of-sample observations taken from the closest area if first iteration for the calculation of calibration weights fails. Only needed if aggData = TRUE with defaults to 0.

w_min

Minimal number of covariates from which informative weights are calculated. Only needed if aggData = TRUE. Defaults to 3.

B

Number of bootstrap replications for MSE estimation procedures. Defaults to 100.

B_adj

Number of bootstrap replications for the adjustment of residual variance proposed by Mendez and Lohr (2001). Defaults to 100.

B_MC

Number of bootstrap populations in the MC version for point estimates of (nonlinear) indicators. Defaults to 100.

threshold

Set a custom threshold for indicators, such as the head count ratio. The threshold can be a known numeric value or function of Y. If the threshold is NULL, 60 % of the median of Y is taken as threshold. Defaults to NULL.

custom_indicator

A list of additional functions containing the indicators to be calculated. These functions must only depend on the target variable Y and optionally the threshold. Defaults to NULL.

initialRandomEffects

Numeric value or vector of initial estimates of random effects. Defaults to 0.

ErrorTolerance

Numeric value to monitor the MERF algorithm's convergence. Defaults to 1e-04.

MaxIterations

Numeric value specifying the maximal amount of iterations for the MERF algorithm. Defaults to 25.

na.rm

Logical. Whether missing values should be removed. Defaults to TRUE.

...

Additional parameters are directly passed to the random forest ranger. Most important parameters are for instance mtry (number of variables to possibly split at in each node), or num.trees (number of trees). For further details on possible parameters see ranger and the example below.

Value

An object of class SAEforest includes point estimates for disaggregated indicators as well as information on the MERF-model. Optionally corresponding MSE estimates are returned. Several generic functions have methods for the returned object of class SAEforest. For a full list and explanation of components and possibilities for objects of class SAEforest, see SAEforestObject.

Details

MERFs combine advantages of regression forests (such as implicit model-selection and robustness properties) with the ability to model hierarchical dependencies.

The MERF algorithm iteratively optimizes two separate steps: a) the random forest function, assuming the random effects term to be correct and b) estimates the random effects part, assuming the OOB-predictions from the forest to be correct. Overall convergence of the algorithm is monitored by log-likelihood of a joint model of both components. For further details see Krennmair and Schmid (2022).

Users that are only interested in the estimation of domain-level means should set meanOnly = TRUE. The MERF requires covariate micro-data. This function, however also allows for the use of aggregated covariate information, by setting aggData = TRUE. Aggregated covariate information is adaptively incorporated through calibration-weights based on empirical likelihood for the estimation of area-level means. See methodological details in Krennmair et al. (2022a)

For the estimation of (nonlinear) poverty indicators and/or quantiles, we need information on the area-specific cumulative distribution function (CDF) of Y. Krennmair et al. (2022b) propose a smearing approach originated by Duan (1983). Alternatively, Monte-Carlo methods are used to simulate the domain-specific CDF of Y.

For the estimation of the MSE, the bootstrap population is built based on a bias-corrected residual variance as discussed in Krennmair and Schmid (2022). The bootstrap bias correction follows Mendez and Lohr (2011).

Note that the MERFmodel object is a composition of elements from a random forest of class ranger and a random effects model of class merMod. Thus, all generic functions are applicable to corresponding objects. For further details on generic functions see ranger and lmer as well as the examples below.

References

Duan, N. (1983). Smearing Estimate: A Nonparametric Retransformation Method. Journal of the American Statistical Association, 78(383), 605–610.

Krennmair, P., & Schmid, T. (2022). Flexible Domain Prediction Using Mixed Effects Random Forests. Journal of Royal Statistical Society: Series C (Applied Statistics) (forthcoming).

Krennmair, P., & Würz, N. & Schmid, T. (2022a). Analysing Opportunity Cost of Care Work using Mixed Effects Random Forests under Aggregated Census Data.

Krennmair, P., & Schmid, T & Tzavidis, Nikos. (2022b). The Estimation of Poverty Indicators Using Mixed Effects Random Forests. Working Paper.

Mendez, G., & Lohr, S. (2011). Estimating residual variance in random forest regression. Computational Statistics & Data Analysis, 55 (11), 2937–2950.

See also

Examples

# \donttest{
# Loading data
data("eusilcA_pop")
data("eusilcA_smp")

income <- eusilcA_smp$eqIncome
X_covar <- eusilcA_smp[,-c(1, 16, 17, 18)]

# Example 1:
# Calculating point estimates and discussing basic generic functions

model1 <- SAEforest_model(Y = income, X = X_covar, dName = "district",
                          smp_data = eusilcA_smp, pop_data = eusilcA_pop)

# SAEforest generics:
summary(model1)
#> ________________________________________________________________
#> Mixed Effects Random Forest for Small Area Estimation
#> ________________________________________________________________
#> Call:
#> SAEforest_model(Y = income, X = X_covar, dName = "district", 
#>     smp_data = eusilcA_smp, pop_data = eusilcA_pop)
#> 
#> Domains
#> ________________________________________________________________
#>  In-sample Out-of-sample Total
#>         70            24    94
#> 
#> Totals:
#> Units in sample: 1945 
#> Units in population: 25000 
#> 
#>                    Min. 1st Qu. Median      Mean 3rd Qu. Max.
#> Sample_domains       14    17.0   22.5  27.78571   29.00  200
#> Population_domains    5   126.5  181.5 265.95745  265.75 5857
#> 
#> Random forest component: 
#> ________________________________________________________________
#>                                            
#> Type:                            Regression
#> Number of trees:                        500
#> Number of independent variables:         14
#> Mtry:                                     3
#> Minimal node size:                        5
#> Variable importance mode:          impurity
#> Splitrule:                         variance
#> Rsquared (OOB):                     0.62041
#> 
#> Structural component of random effects:
#> ________________________________________________________________
#> Linear mixed model fit by maximum likelihood  ['lmerMod']
#> Formula: Target ~ -1 + (1 | district)
#>    Data: data
#>  Offset: forest_preds
#> 
#>      AIC      BIC   logLik deviance df.resid 
#>  39227.0  39238.2 -19611.5  39223.0     1943 
#> 
#> Scaled residuals: 
#>     Min      1Q  Median      3Q     Max 
#> -2.8818 -0.5264 -0.0603  0.4352 11.8525 
#> 
#> Random effects:
#>  Groups   Name        Variance Std.Dev.
#>  district (Intercept) 12594647 3549    
#>  Residual             30763621 5546    
#> Number of obs: 1945, groups:  district, 70
#> 
#> ICC:  0.2904785 
#> 
#> Convergence of MERF algorithm: 
#> ________________________________________________________________
#> Convergence achieved after 10 iterations.
#> A maximum of 25 iterations used and tolerance set to: 1e-04 
#> 
#> Monitored Log-Likelihood:                                                                      
#>  0 -19523.2 -19568.4 -19592.05 -19602.64 -19589.11 -19598.88 -19611.78
#>                               
#>  -19603.29 -19612.73 -19611.52

# Example 2:
# Calculating point + MSE estimates for aggregated covariate data and passing
# arguments to the random forest.
# Note that B is unrealistically low to improve example speed

# remove factor for gender
X_covar <- X_covar[,-1]
model2 <- SAEforest_model(Y = income, X = X_covar, dName = "district",
                          smp_data = eusilcA_smp, pop_data = eusilcA_popAgg,
                          MSE = "nonparametric", popnsize = popNsize,B = 5, mtry = 5,
                          num.trees = 100, aggData = TRUE)
#> Calculation of weights failed for area: Eisenstadt (Stadt)
#> Calculation of weights failed for area: Melk
#> Calculation of weights failed for area: Rust (Stadt)
#> Calculation of weights failed for area: Schärding
#> Calculation of weights failed for area: Villach (Stadt)
#> Calculation of weights failed for area: Wels (Stadt)
#> Calculation of weights failed for area: Wiener Neustadt (Stadt)
#> Error SD Bootstrap started:
#> Bootstrap with 5 rounds started

# SAEforest generics:
summary(model2)
#> ________________________________________________________________
#> Mixed Effects Random Forest for Small Area Estimation
#> ________________________________________________________________
#> Call:
#> SAEforest_model(Y = income, X = X_covar, dName = "district", 
#>     smp_data = eusilcA_smp, pop_data = eusilcA_popAgg, MSE = "nonparametric", 
#>     aggData = TRUE, popnsize = popNsize, B = 5, mtry = 5, num.trees = 100)
#> 
#> Domains
#> ________________________________________________________________
#>  In-sample Out-of-sample Total
#>         70            24    94
#> 
#> Totals:
#> Units in sample: 1945 
#> 
#>                Min. 1st Qu. Median     Mean 3rd Qu. Max.
#> Sample_domains   14      17   22.5 27.78571      29  200
#> 
#> Random forest component: 
#> ________________________________________________________________
#>                                            
#> Type:                            Regression
#> Number of trees:                        100
#> Number of independent variables:         13
#> Mtry:                                     5
#> Minimal node size:                        5
#> Variable importance mode:          impurity
#> Splitrule:                         variance
#> Rsquared (OOB):                     0.65608
#> 
#> Structural component of random effects:
#> ________________________________________________________________
#> Linear mixed model fit by maximum likelihood  ['lmerMod']
#> Formula: Target ~ -1 + (1 | district)
#>    Data: data
#>  Offset: forest_preds
#> 
#>      AIC      BIC   logLik deviance df.resid 
#>  39129.4  39140.5 -19562.7  39125.4     1943 
#> 
#> Scaled residuals: 
#>     Min      1Q  Median      3Q     Max 
#> -4.2203 -0.5099 -0.0409  0.4155 12.1010 
#> 
#> Random effects:
#>  Groups   Name        Variance Std.Dev.
#>  district (Intercept)  8447011 2906    
#>  Residual             29595632 5440    
#> Number of obs: 1945, groups:  district, 70
#> 
#> ICC:  0.2220406 
#> 
#> Convergence of MERF algorithm: 
#> ________________________________________________________________
#> Convergence achieved after 5 iterations.
#> A maximum of 25 iterations used and tolerance set to: 1e-04 
#> 
#> Monitored Log-Likelihood:                                                   
#>  0 -19527.9 -19525.41 -19536.88 -19563.47 -19562.69
summarize_indicators(model2, MSE = TRUE, CV = TRUE)
#> Warning: For objects of class SAEforest_mean or SAEforest_meanAGG, only results for the Indicator "Mean" are reported
#> Indicator/s: Mean
#>                         district      Mean   Mean_MSE    Mean_CV
#> 1                      Amstetten 14153.288   463578.7 0.04810657
#> 2                          Baden 21805.591   258126.3 0.02329959
#> 3                        Bludenz 11906.394   661733.0 0.06832209
#> 4                 Braunau am Inn 12814.693   816657.9 0.07051993
#> 5                        Bregenz 33698.995  7223356.8 0.07975402
#> 6            Bruck an der Leitha 22991.716  1735551.5 0.05729904
#> 7             Bruck-Mürzzuschlag 24860.350   957270.5 0.03935592
#> 8               Deutschlandsberg 23785.282  1983429.9 0.05921069
#> 9                       Dornbirn 19164.019   532192.7 0.03806693
#> 10                      Eferding 13659.162 16769909.7 0.29980668
#> 11            Eisenstadt (Stadt) 29608.069  4430198.2 0.07108885
#> 12           Eisenstadt-Umgebung 23110.247 22973516.3 0.20740019
#> 13                     Feldkirch 17657.821   529488.0 0.04120889
#> 14                   Feldkirchen 14376.373 17567118.9 0.29154187
#> 15                     Freistadt 16922.241  3215447.4 0.10596510
#> 16                       Gmunden 19030.034   281146.9 0.02786295
#> 17                         Gmünd 13552.186  1714141.5 0.09660820
#> 18                  Graz (Stadt) 17916.416   134545.6 0.02047310
#> 19                 Graz-Umgebung 35416.931  1041875.4 0.02882020
#> 20                  Grieskirchen 15629.886  2948423.4 0.10985988
#> 21                   Gänserndorf 20185.350   533606.4 0.03618881
#> 22                       Güssing 18275.257  4475267.2 0.11575666
#> 23                       Hallein 17269.699  2394250.5 0.08959835
#> 24          Hartberg-Fürstenfeld 12239.296  3876452.7 0.16086473
#> 25                      Hermagor 12367.253 21536978.3 0.37524866
#> 26                    Hollabrunn 16252.265  1732016.0 0.08097707
#> 27                          Horn 13700.019 11418631.0 0.24665269
#> 28                          Imst 14210.966  1719744.1 0.09228015
#> 29              Innsbruck (Land) 26959.000  1309937.9 0.04245429
#> 30             Innsbruck (Stadt) 17488.166   149167.6 0.02208478
#> 31                   Jennersdorf 12415.503 10654836.0 0.26291120
#> 32        Kirchdorf an der Krems 16191.971  1415948.1 0.07348927
#> 33                     Kitzbühel 13930.121  2031444.6 0.10231696
#> 34             Klagenfurt (Land) 21673.606  1087431.3 0.04811381
#> 35            Klagenfurt (Stadt) 33054.259  2596825.1 0.04875217
#> 36                    Korneuburg 25865.681  2412093.0 0.06004448
#> 37                  Krems (Land) 15075.867  5135499.1 0.15031733
#> 38    Krems an der Donau (Stadt) 18601.421 39296497.3 0.33700070
#> 39                      Kufstein 19728.278  1000276.4 0.05069566
#> 40                       Landeck 13182.990  2958269.2 0.13046828
#> 41                      Leibnitz 15395.573   959281.0 0.06361757
#> 42                        Leoben 19045.793   369061.0 0.03189701
#> 43                         Lienz 11400.154   430833.0 0.05757630
#> 44                        Liezen 12286.255  1178838.2 0.08837055
#> 45                    Lilienfeld 13301.533  5049607.9 0.16893792
#> 46                  Linz (Stadt) 22363.154   143861.5 0.01696052
#> 47                     Linz-Land 26754.773   405713.1 0.02380719
#> 48                   Mattersburg 20595.336  2621891.9 0.07862099
#> 49                          Melk 12967.553   442660.8 0.05130710
#> 50                    Mistelbach 20115.887   980697.1 0.04922982
#> 51                         Murau 12069.161  4185325.9 0.16950698
#> 52                        Murtal 19179.026  2566777.3 0.08353482
#> 53                       Mödling 41399.246   862865.7 0.02243774
#> 54                   Neunkirchen 15352.054   639222.5 0.05207863
#> 55               Neusiedl am See 18634.918   617594.0 0.04217199
#> 56                Oberpullendorf 17351.933  3245501.0 0.10382286
#> 57                      Oberwart 12253.589   726793.0 0.06957321
#> 58                          Perg 19173.632  1103029.0 0.05477584
#> 59                        Reutte 18675.541   860577.7 0.04967317
#> 60              Ried im Innkreis 13945.013  1644464.9 0.09195882
#> 61                      Rohrbach 13005.926   491439.6 0.05390062
#> 62                  Rust (Stadt) 16300.101  9924437.5 0.19326921
#> 63              Salzburg (Stadt) 20327.806   441754.4 0.03269640
#> 64             Salzburg-Umgebung 29103.602   744519.6 0.02964772
#> 65        Sankt Johann im Pongau 13034.444  2496730.3 0.12122531
#> 66           Sankt Pölten (Land) 17109.056   671323.8 0.04788947
#> 67          Sankt Pölten (Stadt) 17168.734  2646274.6 0.09474999
#> 68        Sankt Veit an der Glan 14887.620  1795806.5 0.09001284
#> 69                      Scheibbs 13424.130  6465492.4 0.18941513
#> 70                        Schwaz 15286.422   601985.3 0.05075596
#> 71                     Schärding 11162.269   651040.9 0.07228558
#> 72           Spittal an der Drau 13134.610   342666.7 0.04456755
#> 73                 Steyr (Stadt) 20295.137 14397523.4 0.18696138
#> 74                    Steyr-Land 25253.106  1217257.0 0.04368942
#> 75              Südoststeiermark 13817.204  3315731.0 0.13178606
#> 76                       Tamsweg 16833.388  5053906.2 0.13354943
#> 77                         Tulln 20998.002   548202.3 0.03526082
#> 78               Urfahr-Umgebung 38841.147  5403933.1 0.05984983
#> 79               Villach (Stadt) 22231.548  2169073.6 0.06624719
#> 80                  Villach Land 18928.555 12697732.6 0.18825461
#> 81                     Voitsberg 17167.095  2106566.1 0.08454555
#> 82                   Vöcklabruck 18208.977   423629.3 0.03574435
#> 83                   Völkermarkt 15316.629  2555153.3 0.10436270
#> 84        Waidhofen an der Thaya 12053.663  8201998.8 0.23759693
#> 85 Waidhofen an der Ybbs (Stadt) 17320.124  2545560.2 0.09211719
#> 86                          Weiz 15266.080   845748.2 0.06024111
#> 87                  Wels (Stadt) 13816.455  1042290.7 0.07389207
#> 88                     Wels-Land 23763.911 13028530.1 0.15189021
#> 89                          Wien 20702.800   467770.1 0.03303598
#> 90        Wiener Neustadt (Land) 20646.361   667884.6 0.03958286
#> 91       Wiener Neustadt (Stadt) 15107.090  3782888.0 0.12874516
#> 92                     Wolfsberg 16842.447  2956297.5 0.10208663
#> 93                   Zell am See  9930.149   936655.1 0.09746172
#> 94                        Zwettl 11753.982 17188541.9 0.35272360

# Example 3:
# Calculating point + MSE estimates and passing arguments to the forest.
# Two additional custom indicators and the threshold is defined as a custom function of Y.
# Note that B is unrealistically low to improve example speed.

model3 <- SAEforest_model(Y = income, X = X_covar, dName = "district", smp_data = eusilcA_smp,
                          pop_data = eusilcA_pop, meanOnly = FALSE, MSE = "nonparametric",
                          B = 5, mtry = 5, num.trees = 100, threshold = function(Y){0.5 *
                          median(Y)}, custom_indicator = list(my_max = function(Y,
                          threshold){max(Y)}, mean40 = function(Y, threshold){
                          mean(Y[Y<=quantile(Y,0.4)])}), smearing = FALSE)
#> Error SD Bootstrap started:
#> MSE Bootstrap with 5 rounds started:

# SAEforest generics:
summary(model3)
#> ________________________________________________________________
#> Mixed Effects Random Forest for Small Area Estimation
#> ________________________________________________________________
#> Call:
#> SAEforest_model(Y = income, X = X_covar, dName = "district", 
#>     smp_data = eusilcA_smp, pop_data = eusilcA_pop, MSE = "nonparametric", 
#>     meanOnly = FALSE, smearing = FALSE, B = 5, threshold = function(Y) {
#>         0.5 * median(Y)
#>     }, custom_indicator = list(my_max = function(Y, threshold) {
#>         max(Y)
#>     }, mean40 = function(Y, threshold) {
#>         mean(Y[Y <= quantile(Y, 0.4)])
#>     }), mtry = 5, num.trees = 100)
#> 
#> Domains
#> ________________________________________________________________
#>  In-sample Out-of-sample Total
#>         70            24    94
#> 
#> Totals:
#> Units in sample: 1945 
#> Units in population: 25000 
#> 
#>                    Min. 1st Qu. Median      Mean 3rd Qu. Max.
#> Sample_domains       14    17.0   22.5  27.78571   29.00  200
#> Population_domains    5   126.5  181.5 265.95745  265.75 5857
#> 
#> Random forest component: 
#> ________________________________________________________________
#>                                            
#> Type:                            Regression
#> Number of trees:                        100
#> Number of independent variables:         13
#> Mtry:                                     5
#> Minimal node size:                        5
#> Variable importance mode:          impurity
#> Splitrule:                         variance
#> Rsquared (OOB):                     0.66844
#> 
#> Structural component of random effects:
#> ________________________________________________________________
#> Linear mixed model fit by maximum likelihood  ['lmerMod']
#> Formula: Target ~ -1 + (1 | district)
#>    Data: data
#>  Offset: forest_preds
#> 
#>      AIC      BIC   logLik deviance df.resid 
#>  39083.5  39094.6 -19539.7  39079.5     1943 
#> 
#> Scaled residuals: 
#>     Min      1Q  Median      3Q     Max 
#> -4.0682 -0.5170 -0.0513  0.4268 12.1747 
#> 
#> Random effects:
#>  Groups   Name        Variance Std.Dev.
#>  district (Intercept)  7689148 2773    
#>  Residual             28970242 5382    
#> Number of obs: 1945, groups:  district, 70
#> 
#> ICC:  0.2097457 
#> 
#> Convergence of MERF algorithm: 
#> ________________________________________________________________
#> Convergence achieved after 4 iterations.
#> A maximum of 25 iterations used and tolerance set to: 1e-04 
#> 
#> Monitored Log-Likelihood:                                          
#>  0 -19519.14 -19541.96 -19537.78 -19539.73
summarize_indicators(model3, MSE = FALSE, CV = TRUE, indicator = c("Gini", "my_max", "mean40"))
#> Indicator/s: Gini
#>  Indicator/s: my_max
#>  Indicator/s: mean40
#>                         district      Gini    Gini_CV    my_max my_max_CV
#> 1                      Amstetten 0.2496166 0.06139391  53367.94 0.5124650
#> 2                          Baden 0.1821222 0.07004379  69365.94 0.3929977
#> 3                        Bludenz 0.2836343 0.10597374  45119.26 0.2913200
#> 4                 Braunau am Inn 0.2769782 0.07798593  53789.77 0.5638769
#> 5                        Bregenz 0.1623321 0.07778439  82887.97 0.1120534
#> 6            Bruck an der Leitha 0.1786187 0.07669723  69086.14 0.1365179
#> 7             Bruck-Mürzzuschlag 0.1668700 0.09485270  71669.91 0.2235948
#> 8               Deutschlandsberg 0.1897958 0.08023022  57235.93 0.1824685
#> 9                       Dornbirn 0.2181016 0.06908320  56893.75 0.3678049
#> 10                      Eferding 0.2345167 0.18489711  42421.92 0.6557794
#> 11            Eisenstadt (Stadt) 0.1910494 0.17611874  68312.35 0.1181623
#> 12           Eisenstadt-Umgebung 0.2009430 0.12954595  58555.45 0.1984942
#> 13                     Feldkirch 0.2317475 0.10229294  59428.59 0.3881409
#> 14                   Feldkirchen 0.2446526 0.18556621  40754.56 0.2536565
#> 15                     Freistadt 0.2154424 0.12854458  53101.29 0.4965489
#> 16                       Gmunden 0.1999884 0.09455086  61906.52 0.2955842
#> 17                         Gmünd 0.2561330 0.29310131  39789.79 0.5867452
#> 18                  Graz (Stadt) 0.2176966 0.06736820  76946.33 0.1984235
#> 19                 Graz-Umgebung 0.1861077 0.16841540  95132.01 0.2606042
#> 20                  Grieskirchen 0.2167425 0.10161904  52493.09 0.2587489
#> 21                   Gänserndorf 0.1995323 0.10132979  61202.70 0.4011403
#> 22                       Güssing 0.2331344 0.08623459  42441.73 0.3973735
#> 23                       Hallein 0.2258108 0.12961906  51111.93 0.3519029
#> 24          Hartberg-Fürstenfeld 0.2645503 0.07211716  55098.67 0.3520621
#> 25                      Hermagor 0.2523298 0.16283129  33864.60 0.7615479
#> 26                    Hollabrunn 0.2223246 0.11138456  48751.66 0.4628220
#> 27                          Horn 0.2338976 0.19230341  43554.09 0.3160812
#> 28                          Imst 0.2377699 0.08142345  46086.11 0.2619329
#> 29              Innsbruck (Land) 0.2040890 0.08602366  77048.75 0.1486210
#> 30             Innsbruck (Stadt) 0.2065810 0.10692276  60773.36 0.4932663
#> 31                   Jennersdorf 0.2494483 0.17889506  34599.53 0.1609961
#> 32        Kirchdorf an der Krems 0.2325506 0.09950261  50549.39 0.5055384
#> 33                     Kitzbühel 0.2505483 0.11280844  45226.13 0.4784815
#> 34             Klagenfurt (Land) 0.1825540 0.08386088  57705.76 0.3781472
#> 35            Klagenfurt (Stadt) 0.1639245 0.06938104  81294.83 0.3479100
#> 36                    Korneuburg 0.1603203 0.08917155  63771.42 0.3168946
#> 37                  Krems (Land) 0.2391255 0.08436111  50278.43 0.2209006
#> 38    Krems an der Donau (Stadt) 0.2199381 0.11800657  40318.15 0.3457054
#> 39                      Kufstein 0.1847680 0.07407301  58083.20 0.2491800
#> 40                       Landeck 0.2615573 0.07132085  38413.82 0.7690219
#> 41                      Leibnitz 0.2315831 0.11104392  53370.41 0.4808914
#> 42                        Leoben 0.1927015 0.12717968  58604.30 0.2134696
#> 43                         Lienz 0.2608947 0.12025133  45782.32 0.4455413
#> 44                        Liezen 0.2734132 0.13385868  53213.01 0.4873165
#> 45                    Lilienfeld 0.2443897 0.19718093  42680.27 0.7047444
#> 46                  Linz (Stadt) 0.1866490 0.07613662  76815.99 0.1997204
#> 47                     Linz-Land 0.1685493 0.04476869  77998.89 0.2628937
#> 48                   Mattersburg 0.2068680 0.18340054  52645.18 0.2048766
#> 49                          Melk 0.2735880 0.08773104  48230.46 0.2642358
#> 50                    Mistelbach 0.2036992 0.12599481  60321.48 0.1358387
#> 51                         Murau 0.2529370 0.14983048  35311.93 0.3188117
#> 52                        Murtal 0.2150873 0.07773061  53926.56 0.4617912
#> 53                       Mödling 0.1496726 0.19444333  87746.94 0.1476431
#> 54                   Neunkirchen 0.2167925 0.13399309  56766.39 0.3887731
#> 55               Neusiedl am See 0.2056638 0.09013821  54394.24 0.2373192
#> 56                Oberpullendorf 0.2331675 0.21982944  46670.48 0.5692763
#> 57                      Oberwart 0.2583104 0.14040406  47608.62 0.5629113
#> 58                          Perg 0.2127375 0.09274623  53158.31 0.1649124
#> 59                        Reutte 0.1984683 0.10557539  43351.78 0.2284484
#> 60              Ried im Innkreis 0.2474063 0.12518031  50045.91 0.3044791
#> 61                      Rohrbach 0.2188741 0.12589123  48102.63 0.2523869
#> 62                  Rust (Stadt) 0.2697352 0.18642302  23357.17 0.1908323
#> 63              Salzburg (Stadt) 0.1932685 0.07948163  68674.85 0.1591554
#> 64             Salzburg-Umgebung 0.1679929 0.09598427  89310.34 0.1694144
#> 65        Sankt Johann im Pongau 0.2453427 0.09941047  53633.28 0.2670028
#> 66           Sankt Pölten (Land) 0.2298954 0.08762768  62305.47 0.2294852
#> 67          Sankt Pölten (Stadt) 0.2221361 0.16899913  49775.06 0.5028214
#> 68        Sankt Veit an der Glan 0.2455244 0.10072608  51492.05 0.3967974
#> 69                      Scheibbs 0.2480215 0.19712089  42353.38 0.3760305
#> 70                        Schwaz 0.2298244 0.15086416  52413.72 0.2150530
#> 71                     Schärding 0.3029931 0.08659790  43510.19 0.3755736
#> 72           Spittal an der Drau 0.2589875 0.07263732  58545.46 0.2843762
#> 73                 Steyr (Stadt) 0.2144196 0.28573383  47224.17 0.5252292
#> 74                    Steyr-Land 0.1730117 0.09023882  57873.40 0.4890363
#> 75              Südoststeiermark 0.2509470 0.06872722  55264.52 0.4612891
#> 76                       Tamsweg 0.2162900 0.10130196  40879.56 0.1156401
#> 77                         Tulln 0.2054317 0.08278396  61652.68 0.1982365
#> 78               Urfahr-Umgebung 0.1443740 0.15581748  89357.50 0.1001912
#> 79               Villach (Stadt) 0.1941162 0.11158736  60944.94 0.1558012
#> 80                  Villach Land 0.2069803 0.07776880  56018.43 0.2372981
#> 81                     Voitsberg 0.2100369 0.12259307  48686.29 0.5232139
#> 82                   Vöcklabruck 0.2111293 0.05842563  63596.83 0.2556668
#> 83                   Völkermarkt 0.2349521 0.10301081  47016.00 0.2725515
#> 84        Waidhofen an der Thaya 0.2660027 0.15929675  35849.69 0.1811430
#> 85 Waidhofen an der Ybbs (Stadt) 0.2109265 0.26315102  35962.63 0.1605012
#> 86                          Weiz 0.2452848 0.08436544  51603.52 0.3177342
#> 87                  Wels (Stadt) 0.2408120 0.09499697  47470.60 0.2739218
#> 88                     Wels-Land 0.2001741 0.08472676  59415.56 0.2160322
#> 89                          Wien 0.2447866 0.05381822 109676.70 0.2103532
#> 90        Wiener Neustadt (Land) 0.2057335 0.09337493  57415.41 0.2618388
#> 91       Wiener Neustadt (Stadt) 0.2105591 0.09349449  45909.63 0.2722959
#> 92                     Wolfsberg 0.2332638 0.09679308  50785.10 0.2186416
#> 93                   Zell am See 0.2933377 0.02306185  45910.49 0.3219887
#> 94                        Zwettl 0.2677241 0.11152893  45474.82 0.5364367
#>       mean40  mean40_CV
#> 1   8104.170 0.02902325
#> 2  15735.431 0.02839009
#> 3   6234.411 0.22953106
#> 4   6253.489 0.25399596
#> 5  23840.219 0.03849579
#> 6  16660.491 0.03372998
#> 7  17345.005 0.09502401
#> 8  14506.407 0.11582412
#> 9  12450.326 0.09412046
#> 10  9565.550 0.35804410
#> 11 27297.832 0.30017683
#> 12 17758.733 0.20886305
#> 13 10075.268 0.13719385
#> 14  8402.480 0.27306219
#> 15 10754.675 0.10974376
#> 16 13368.639 0.04987205
#> 17  7485.412 0.85402883
#> 18 11365.140 0.05354266
#> 19 24264.907 0.05996671
#> 20 10584.374 0.08537967
#> 21 13424.339 0.10829529
#> 22 10275.955 0.14684277
#> 23 10024.918 0.10733773
#> 24  7309.488 0.15066485
#> 25  7087.851 0.31167004
#> 26 10231.910 0.09760702
#> 27  9405.537 0.44537825
#> 28  8610.950 0.15825954
#> 29 17869.057 0.05511827
#> 30 11623.388 0.07832363
#> 31  7667.007 0.23136868
#> 32 10030.946 0.12467688
#> 33  7138.563 0.10963907
#> 34 15789.840 0.09034039
#> 35 24134.944 0.05707198
#> 36 20230.548 0.06131129
#> 37  9220.980 0.12094249
#> 38 10874.673 0.11124036
#> 39 14157.177 0.05572605
#> 40  6690.407 0.22772454
#> 41  9564.511 0.08729996
#> 42 13783.134 0.10627174
#> 43  7211.216 0.14491196
#> 44  6485.282 0.23510147
#> 45  8939.741 0.48175894
#> 46 15460.731 0.06819121
#> 47 19818.833 0.03775992
#> 48 14359.113 0.36398719
#> 49  6209.175 0.11881243
#> 50 13183.924 0.07627689
#> 51  7260.136 0.21362813
#> 52 12128.811 0.12253665
#> 53 30477.641 0.04364185
#> 54 10333.526 0.14205719
#> 55 12630.609 0.06372049
#> 56 10385.815 0.41585492
#> 57  7387.320 0.22284874
#> 58 11466.888 0.14186941
#> 59 12948.996 0.13908392
#> 60  8129.388 0.25564883
#> 61  9546.130 0.12819871
#> 62  6752.316 0.21863140
#> 63 13423.428 0.05639137
#> 64 22100.403 0.04523701
#> 65  8542.659 0.13930779
#> 66 10368.832 0.06807740
#> 67 10210.859 0.17509699
#> 68  8725.549 0.13554502
#> 69  7794.362 0.21527478
#> 70  9618.193 0.09631738
#> 71  4738.325 0.29389391
#> 72  7703.324 0.09920418
#> 73 12093.106 0.43165920
#> 74 17657.622 0.09946456
#> 75  7923.985 0.21775466
#> 76 10633.185 0.12228878
#> 77 13220.184 0.07271324
#> 78 32063.473 0.05584124
#> 79 15122.046 0.09816730
#> 80 12793.717 0.04483305
#> 81 11390.310 0.15038505
#> 82 11800.726 0.06066700
#> 83  9235.364 0.13208011
#> 84  6855.515 0.33219771
#> 85 11547.659 0.45972960
#> 86  8482.244 0.08870874
#> 87  8464.953 0.09782143
#> 88 13509.826 0.08020484
#> 89 12093.349 0.02981109
#> 90 12484.619 0.08173127
#> 91 11126.650 0.11644553
#> 92 10536.540 0.09727307
#> 93  4909.782 0.13866430
#> 94  7230.085 0.23743497
# }